feat: switch tests to esm
